test(agents): split model-fallback e2e store helpers into test support
The embedded-runner fallback e2e file crossed the max-lines lint cap;
the mock-free config/workspace/auth-store helpers move to a colocated
test-support module with grep-unique exported names.
